Introduced new Main
Main was changed to SearchRunner and a new Main was introduced to drive
the application. MainApplet now runs off of this Main.
