+++ routeFinder/control/.gitignore
@@ -1,1 +1,2 @@
 /Main.class
+/SearchRunner.class
