+++ routeFinder/view/MainApplet.java
@@ -1,45 +1,45 @@
 package routeFinder.view;
 
 import processing.core.*;
-import routeFinder.control.Main;
+import processing.data.XML;
+import routeFinder.control.SearchRunner;
 
 /*
  * Original implementation by Gary Dahl
  */
-public class MainApplet extends PApplet {
+public class MainApplet {
 	private MapView mapView; // View component to draw map to the applet
-	private Main main;
+	private SearchRunner main;
 	
-	private String mapFileName = "map.osm";
+	private XML mapData;
+	private PApplet applet;
 
 	private final float MAP_HEIGHT_RATIO = 0.8f;
 	private final float MAP_WIDTH_RATIO = 1.0f;
 
 	private float MAP_TOP;
 	private float MAP_BOTTOM;
-
-	// initialize window
-	public void settings() {
-		size(800, 600);
+	
+	public MainApplet(PApplet applet, XML mapData) {
+		this.applet = applet;
+		this.mapData = mapData;
+		setup();
 	}
 
 	// load map, and initialize fields along with processing modes
 	public void setup() {
-		MAP_TOP = this.height * ( 1 - MAP_HEIGHT_RATIO ) / 2;
-		MAP_BOTTOM = this.height * ( MAP_HEIGHT_RATIO + ( 1 - MAP_HEIGHT_RATIO ) / 2 );
+		MAP_TOP = applet.height * ( 1 - MAP_HEIGHT_RATIO ) / 2;
+		MAP_BOTTOM = applet.height * ( MAP_HEIGHT_RATIO + ( 1 - MAP_HEIGHT_RATIO ) / 2 );
 		
-		int mapWidth = Math.round(this.width * MAP_WIDTH_RATIO);
-		int mapHeight = Math.round(this.height * MAP_HEIGHT_RATIO);
+		int mapWidth = Math.round(applet.width * MAP_WIDTH_RATIO);
+		int mapHeight = Math.round(applet.height * MAP_HEIGHT_RATIO);
 		
-		main = new Main(this);
+		main = new SearchRunner(applet);
 		
-		main.openMap(loadXML(mapFileName), mapWidth, mapHeight);
+		main.openMap(mapData, mapWidth, mapHeight);
 		main.initializeGuiPoints();
 		
-		mapView = new MapView(main.map, this);
-
-		textAlign(CENTER);
-		rectMode(CORNER);
+		mapView = new MapView(main.map, applet);
 	}
 
 	// update
@@ -67,7 +67,7 @@ public class MainApplet extends PApplet {
 	 * Clears the entire applet background with blue and draws the map on top.
 	 */
 	private void drawMap() {
-		background(0, 0, 127); // clear display
+		applet.background(0, 0, 127); // clear display
 		mapView.drawMap();
 	}
 
@@ -83,18 +83,18 @@ public class MainApplet extends PApplet {
 	 * Clears the top of the applet with black.
 	 */
 	private void drawTopPane() {
-		stroke(0);
-		fill(0);
-		rect(0, 0, width, MAP_TOP);
+		applet.stroke(0);
+		applet.fill(0);
+		applet.rect(0, 0, applet.width, MAP_TOP);
 	}
 
 	/*
 	 * Clears the bottom of the applet with black.
 	 */
 	private void drawBottomPane() {
-		stroke(0);
-		fill(0);
-		rect(0, MAP_BOTTOM, width, height);
+		applet.stroke(0);
+		applet.fill(0);
+		applet.rect(0, MAP_BOTTOM, applet.width, applet.height);
 	}
 
 	/*
@@ -111,10 +111,10 @@ public class MainApplet extends PApplet {
 	 * the frontier, and how many there are in the top bar.
 	 */
 	private void drawFrontierProgress() {
-		stroke(color(127, 127, 0));
-		fill(color(255, 255, 0));
+		applet.stroke(applet.color(127, 127, 0));
+		applet.fill(applet.color(255, 255, 0));
 		mapView.drawPoints(main.frontierPoints());
-		text("FRONTIER: " + main.frontierPoints().size(), width / 2, 16);
+		applet.text("FRONTIER: " + main.frontierPoints().size(), applet.width / 2, 16);
 	}
 
 	/*
@@ -122,10 +122,10 @@ public class MainApplet extends PApplet {
 	 * main.searched, and the number of points explored in the top bar
 	 */
 	private void drawExploredProgress() {
-		stroke(color(127, 0, 0));
-		fill(color(255, 0, 0));
+		applet.stroke(applet.color(127, 0, 0));
+		applet.fill(applet.color(255, 0, 0));
 		mapView.drawPoints(main.exploredPoints());
-		text("EXPLORED: " + main.exploredPoints().size(), width / 2, 32);
+		applet.text("EXPLORED: " + main.exploredPoints().size(), applet.width / 2, 32);
 	}
 
 	/*
@@ -144,16 +144,16 @@ public class MainApplet extends PApplet {
 	 * Draws the instructions for beginning a new main.search.
 	 */
 	private void drawPromptToComputeANewSolution() {
-		fill(255);
-		text("Press <0>, <1>, or <2> to find a path from the green to red circle.", width / 2, height - 32);
+		applet.fill(255);
+		applet.text("Press <0>, <1>, or <2> to find a path from the green to red circle.", applet.width / 2, applet.height - 32);
 	}
 
 	/*
 	 * Draws the instructions to continue in a main.search.
 	 */
 	private void drawPromptToContinueSearch() {
-		fill(255); // display prompt to continue exploring or reset main.search
-		text("Press <Enter> to continue or <spacebard> to step through main.search.", width / 2, height - 32);
+		applet.fill(255); // display prompt to continue exploring or reset main.search
+		applet.text("Press <Enter> to continue or <spacebard> to step through main.search.", applet.width / 2, applet.height - 32);
 	}
 
 	/*
@@ -161,9 +161,5 @@ public class MainApplet extends PApplet {
 	 */
 	private void drawSolution() {
 		mapView.drawRoute(main.searchSolution());
-	}
-
-	public static void main(String args[]) {
-		PApplet.main(new String[] { "routeFinder.view.MainApplet" });
 	}
 }
