+++ tests/ApplicationRunner.java
@@ -2,16 +2,16 @@ package tests;
 
 import static org.junit.Assert.*;
 
-import routeFinder.control.Main;
+import routeFinder.control.SearchRunner;
 import routeFinder.model.Point;
 import processing.core.PApplet;
 
 public class ApplicationRunner {
 
-	private Main main;
+	private SearchRunner main;
 	
 	public void start() {
-		main = new Main();
+		main = new SearchRunner();
 	}
 	
 	public void openMap(String mapFileName) {
